pyproject.toml
[project]
name = "epicevents"
version = "0.1.0"
description = ""
authors = [
    {name = "Charles DZADU",email = "dzaducharles@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlmodel>=0.0.14",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pyjwt>=2.8.0",
    "argon2-cffi>=23.1.0",
    "sentry-sdk>=1.40.0",
]

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-cov = ">=4.1.0"
black = ">=23.12.0"
flake8 = ">=7.0.0"
mypy = ">=1.8.0"

[project.scripts]
epicevents = "epicevents.cli.main:cli"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "build", "dist"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["epicevents/tests"]
python_files = ["test_*.py"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
